MODFIN-82 updated module descriptor
