+++ src/main/java/org/folio/rest/helper/TransactionsHelper.java
@@ -9,11 +9,14 @@ import static org.folio.rest.util.ResourcePathResolver.resourcesPath;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
+import org.folio.rest.jaxrs.model.AwaitingPayment;
+import org.folio.rest.jaxrs.model.Encumbrance;
 import org.folio.rest.jaxrs.model.Transaction;
 import org.folio.rest.jaxrs.model.TransactionCollection;
 
 import io.vertx.core.Context;
 import me.escoffier.vertx.completablefuture.VertxCompletableFuture;
+import org.folio.rest.util.MoneyUtils;
 
 public class TransactionsHelper extends AbstractHelper {
 
@@ -40,6 +43,31 @@ public class TransactionsHelper extends AbstractHelper {
 
   public CompletableFuture<Void> updateTransaction(Transaction transaction) {
     return handleUpdateRequest(resourceByIdPath(TRANSACTIONS, transaction.getId(), lang), transaction);
+  }
+
+  /**
+   * Get the {@link Transaction} (encumbrance) from storage and update the encumbered / awaiting payment amounts
+   *
+   * @param awaitingPayment {@link AwaitingPayment} object
+   * @return {@link CompletableFuture<Void>} returns empty result
+   */
+  public CompletableFuture<Void> moveToAwaitingPayment(AwaitingPayment awaitingPayment) {
+    TransactionsHelper transactionsHelper = new TransactionsHelper(okapiHeaders, ctx, lang);
+
+    return transactionsHelper.getTransaction(awaitingPayment.getEncumbranceId())
+      .thenApply(tr -> modifyTransaction(tr, awaitingPayment))
+      .thenCompose(transactionsHelper::updateTransaction);
+  }
+
+  private Transaction modifyTransaction(Transaction transaction, AwaitingPayment awaitingPayment) {
+    Double currentAwaitingPaymentAmount = transaction.getEncumbrance().getAmountAwaitingPayment();
+    String currency = transaction.getCurrency();
+
+    transaction.getEncumbrance()
+      .setAmountAwaitingPayment(MoneyUtils.sumDoubleValues(currentAwaitingPaymentAmount, awaitingPayment.getAmountAwaitingPayment(), currency));
+
+    transaction.getEncumbrance().setStatus(awaitingPayment.getReleaseEncumbrance() ? Encumbrance.Status.RELEASED : Encumbrance.Status.UNRELEASED);
+    return transaction;
   }
 
 }
